Moved files to respective core/viewModels/views folders.
